# app_backend/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gestión de Inventario, Préstamos y Penalizaciones
  description: >
    Esta API maneja la autenticación de usuarios, administración de perfiles, 
    control centralizado de stock (materiales, herramientas, placas, libros), 
    flujo de préstamos/reservas mediante validación por QR, gestión de penalizaciones 
    y generación de reportes/estadísticas.
  version: 2.0.0
servers:
  - url: /
    description: Servidor Base Local
paths:
...
  /api/usuarios:
    get:
      tags:
        - Usuarios
      summary: Lista los usuarios registrados
      description: Devuelve una lista completa de usuarios. (Solo Admin).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuarios obtenida con éxito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDetail'
        '403':
          description: Prohibido. Permisos insuficientes.

    post:
      tags:
        - Usuarios
      summary: Alta (ABM) de un nuevo usuario
      description: Permite registrar un nuevo usuario en el sistema. (Solo Admin).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuario creado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '400':
          description: Datos inválidos o conflicto.
        '403':
          description: Prohibido. Permisos insuficientes.
...
  /api/usuario/{usuario_id}:
    parameters:
      - name: usuario_id
        in: path
        required: true
        description: Identificador único del usuario.
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Usuarios
      summary: Obtiene un usuario por ID (endpoint de compatibilidad)
      description: Endpoint agregado por compatibilidad. Consulta usuarios en la tabla usuario.
      responses:
        '200':
          description: Usuario encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '400':
          description: ID inválido.
        '404':
          description: Usuario no encontrado.
  /api/usuarios/{usuario_id}:
    parameters:
      - name: usuario_id
        in: path
        required: true
        description: Identificador único del usuario.
        schema:
          type: integer
          format: int64
    get:
      tags:
        - Usuarios
      summary: Obtiene los detalles de un usuario específico
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Detalles del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '404':
          description: Usuario no encontrado.

    put:
      tags:
        - Usuarios
      summary: Modifica los datos de un usuario existente
      description: Reemplazo completo de la información general del usuario.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuario modificado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '400':
          description: Datos de actualización inválidos.
        '404':
          description: Usuario no encontrado.

    delete:
      tags:
        - Usuarios
      summary: Baja o desactivación lógica de un usuario (Opción 1)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Usuario dado de baja con éxito.
        '404':
          description: Usuario no encontrado.
...
  /api/usuarios/{usuario_id}/reservas:
    get:
      tags:
        - Usuarios
      summary: Obtiene el historial personal de préstamos/pedidos
      description: Admin y bibliotecario pueden consultar cualquier usuario; alumno y profesor solo pueden consultar su propio historial.
      parameters:
        - name: usuario_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Historial obtenido correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanCompact'
        '404':
          description: Alumno no encontrado.
        '403':
          description: Prohibido. El usuario autenticado no puede consultar reservas ajenas.


  # ==========================================
  # 3. Inventario y Materiales (/api/items)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
          format: email
        puntaje:
          type: integer
        rol:
          type: string
          enum: [alumno, profesor, bibliotecario, admin]
        carrera:
          type: string
          nullable: true
...
    UserCreate:
      type: object
      required:
        - nombre
        - email
        - rol
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          type: string
          enum: [alumno, profesor, bibliotecario, admin]
        carrera:
          type: string
        puntaje:
          type: integer
    UserUpdate:
      type: object
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        rol:
          type: string
          enum: [alumno, profesor, bibliotecario, admin]
        carrera:
          type: string
        puntaje:
          type: integer
...
    LoanCompact:
      type: object
      properties:
        id:
          type: integer
          format: int64
        id_reservado:
          type: integer
          format: int64
        nombre_art:
          type: string
        estado_reserva:
          type: string
        fecha_retiro:
          type: string
          format: date-time
...